Uses summary as an exploration in the benchmark
When scores are too close to the threshold, there is an option to try
to re-calculate the scores using a summary of the incoming Telegram
message instead of the real message.

In theory this would help dealing with different vocabulary used in the
spam message. The score after summarizing should:
- be smaller than the one using the full message for non-spam messages
- by higher than the one using the full message for spam messages

No labels reflected that in the benchmarks I ran locally, so this is
committed as an exploration tool and not as a feature at this point.